% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AA_knowledge_test}
\alias{AA_knowledge_test}
\title{Example of a test to showcase the Adaptive Assessment tools}
\format{
A assessment object based on six multiple choice items.
\describe{
\item{states}{Structure with six items and twelve states.}
\item{likelihood}{Uniformly distributed across the states.}
\item{Stopping_rule}{The stopping rules is "likelihood_maximization" with a criterion of .5.}
\item{Question}{Six multiple choice items.}
\item{adaptive}{The value is set on TRUE allowing adaptive administration.}
\item{repetition}{The value is set on FALSE not allowing items repetition.}
}
}
\source{
{Created to serve as an Example for this package.}
}
\usage{
AA_knowledge_test
}
\description{
The dataset contains a six items test concerning basic knowledge on computerized adaptive assessment.
This test has been developed during the creation of this R package.
}
\examples{
data(AA_knowledge_test) #Lazy loading
}
\keyword{datasets}
