% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.assessment}
\alias{plot.assessment}
\title{Plot (Method)}
\usage{
\method{plot}{assessment}(x, bg_color = NULL, verices_color = "black", ...)
}
\arguments{
\item{x}{Assessment object to display.}

\item{bg_color}{Background color. The default is white.}

\item{verices_color}{Color of the nodes label.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
The function returns the graph representation of the structure.
}
\description{
Function used to plot an assessment object.
}
\examples{
#  Example: Random items presentation

token<- assessment(N_items = 5 ,adaptive = FALSE)
plot(token)
}
