% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Question.R
\name{print_question}
\alias{print_question}
\title{Print the question}
\usage{
print_question(
  TextQuestion,
  TextResponse = list(correct = "correct", incorrect = c("wrong1", "wrong2", "wrong3"))
)
}
\arguments{
\item{TextQuestion}{A character contain the text of the question.}

\item{TextResponse}{A list containing for each question the correct and incorrect answers.}
}
\value{
The answers of the question as numeric (1) if is correct and (0) if is false.
}
\description{
Function to print in the console the item and collect the answer.
}
