% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Assessment.R
\name{run_Assessment}
\alias{run_Assessment}
\title{run Assessment shinyApp}
\usage{
run_Assessment(file = NULL)
}
\arguments{
\item{file}{Logical value.}
}
\value{
The results of the assessment
}
\description{
This function run a shinyApp in which the is available:
iii) The GUI implementation of the adaptive assessment
}
\examples{
# Try the test example yourself
data(AA_knowledge_test)
if(interactive()){
run_Assessment(AA_knowledge_test)}
}
