% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Practice.R
\name{run_Practice}
\alias{run_Practice}
\title{run Practice Assessment shinyApp}
\usage{
run_Practice()
}
\value{
The results of the assessment
}
\description{
This function run a shinyApp in which the following pages are available:
i) a brief introduction of the assessment tools; ii) the GUI to create "assessment" files from a "csv" of items;
iii) the GUI implementation of the adaptive assessment; and iv) the GUI implementation of the outcome.
}
\examples{
# Try to build your test
if(interactive()){
run_Practice()}
}
