% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.assessment}
\alias{summary.assessment}
\title{Summary (Method)}
\usage{
\method{summary}{assessment}(object, ...)
}
\arguments{
\item{object}{Assessment object to display}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
The function return a summary of the information in the a assessment object
}
\description{
Function used to produce result summaries of an assessment object.
}
\examples{
#   Example of random items presentation

token<- assessment(N_items = 5 ,adaptive = FALSE, ki=15)
summary(token)
}
